@@ -35,32 +35,33 @@
     "@angular/common": "^16.1.1",
     "@angular/compiler": "^16.1.1",
     "@angular/core": "^16.1.1",
-    "@angular/forms": "^16.2.6",
-    "@angular/localize": "^16.2.6",
+    "@angular/forms": "^16.2.8",
+    "@angular/localize": "^16.2.8",
     "@angular/material": "^16.1.1",
-    "@angular/material-moment-adapter": "^16.2.5",
+    "@angular/material-moment-adapter": "^16.2.7",
     "@angular/platform-browser": "^16.1.1",
-    "@angular/platform-browser-dynamic": "^16.2.6",
-    "@angular/router": "^16.2.6",
+    "@angular/platform-browser-dynamic": "^16.2.8",
+    "@angular/router": "^16.2.8",
     "@dlr-shepard/shepard-client": "2023.4.11",
-    "@supabase/supabase-js": "^2.36.0",
+    "@supabase/supabase-js": "^2.38.0",
     "@swimlane/ngx-graph": "^8.2.2",
+    "@wogra/wogra-ui-kit": "^0.0.14",
     "buffer": "^6.0.3",
     "cross-env": "7.0.3",
-    "csv-parse": "^5.5.0",
+    "csv-parse": "^5.5.2",
     "jszip": "^3.10.1",
     "lodash": "^4.17.21",
+    "mime": "^4.0.0-beta.1",
     "moment": "^2.29.4",
     "node-sass-json-importer": "^4.3.0",
     "rxjs": "~7.8.0",
-    "sass": "^1.68.0",
-    "snakecase-keys": "^5.4.7",
+    "sass": "^1.69.0",
     "tslib": "^2.6.2",
     "yaml": "^2.3.2",
     "zone.js": "^0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.3",
+    "@angular-devkit/build-angular": "^16.2.5",
     "@angular-eslint/builder": "16.0.3",
     "@angular-eslint/eslint-plugin": "16.0.3",
     "@angular-eslint/eslint-plugin-template": "16.0.3",
@@ -75,7 +76,7 @@
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "cypress": "^12.17.4",
-    "eslint": "^8.50.0",
+    "eslint": "^8.51.0",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jsdoc": "^40.3.0",
