@@ -1,37 +1,51 @@
-import { Component } from '@angular/core';
-import { Breadcrumb } from '../../design/components/molecules/breadcrumb/breadcrumb.component';
+import { Component, OnDestroy } from '@angular/core';
+import { Breadcrumb } from '@wogra/wogra-ui-kit';
 import { yamlToProcess$ } from '../../shared/util/common-utils';
-import { catchError, of, switchMap, take } from 'rxjs';
+import { catchError, of, Subject, switchMap, take, takeUntil } from 'rxjs';
 import { Router } from '@angular/router';
 import { AlertService } from '../../services/alert.service';
 import { BundleService } from '../../services';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { fileExtensionValidator } from '@wogra/wogra-ui-kit';
 
 @Component({
   selector: 'app-bundle-new',
   templateUrl: './bundle-new.component.html',
   styleUrls: ['./bundle-new.component.scss'],
 })
-export class BundleNewComponent {
+export class BundleNewComponent implements OnDestroy {
   public breadcrumbs: Breadcrumb[] = [
     { label: $localize`:@@header.overview:Overview`, link: '/overview' },
     { label: $localize`:@@header.new_bundle:New Bundle` },
   ];
-  public isValid = false;
-  public loading = false;
+  protected loading = false;
+  protected file: File | undefined;
+  protected formGroup: FormGroup = new FormGroup({
+    file: new FormControl('', [
+      Validators.required,
+      fileExtensionValidator(['yaml', 'yml', 'json']),
+    ]),
+  });
+  private _destroy$: Subject<void> = new Subject<void>();
 
   constructor(
     private readonly router: Router,
     private readonly alert: AlertService,
     private readonly bundleService: BundleService
-  ) {}
+  ) {
+    const fileForm = this.formGroup.get('file');
+    fileForm?.valueChanges.pipe(takeUntil(this._destroy$)).subscribe(value => {
+      this.file = value;
+    });
+  }
 
-  public uploadFile(file: File | null): void {
-    if (!file) {
+  public uploadFile(): void {
+    if (!this.file) {
       return;
     }
     this.loading = true;
-    const bundleName = file.name.split('.')[0];
-    yamlToProcess$(file)
+    const bundleName = this.file.name.split('.')[0];
+    yamlToProcess$(this.file)
       .pipe(
         take(1),
         switchMap(procs =>
@@ -53,5 +67,10 @@ export class BundleNewComponent {
 
   public cancel(): void {
     this.router.navigate(['']);
+  }
+
+  public ngOnDestroy(): void {
+    this._destroy$.next();
+    this._destroy$.complete();
   }
 }
